fix clang-tidy complaint about falcon
See https://circleci.com/gh/PQClean/PQClean/7991

+++ crypto_sign/falcon-512/clean/fpr.c
@@ -538,7 +538,7 @@ fpr_expm_p63(fpr x, fpr ccs) {
     };
 
     uint64_t z, y;
-    unsigned u;
+    size_t u;
     uint32_t z0, z1, y0, y1;
     uint64_t a, b;
